Add explicit platform to lcas-jammy docker compose file
On m1 mac, docker tries to start the docker image, which was
originally built for amd64 in native mode and complains that the cpu architecture does not match. Add the platform option allow docker to start the container in emulation mode, which enables the whole thing to run without problem on m1 mac.